LongTensor for clas_tgt

+++ src/losses.py
@@ -91,13 +91,13 @@ class RetinaNetLosses(nn.Module):
         clas_pred = clas_pred[clas_mask]
 
         clas_tgt = clas_tgt + 1
-        # Add background class to account for background in `matches`. When there are no 
+        # Add background class to account for background in `matches`. When there are no
         # matches
         clas_tgt = torch.cat([clas_tgt.new_zeros(1).long(), clas_tgt])
         clas_tgt = clas_tgt[matches[clas_mask]]
-        clas_tgt = F.one_hot(clas_tgt, num_classes=self.n_c + 1)[
-            :, 1:
-        ]  # no loss for the first(background) class
+        clas_tgt = torch.LongTensor(
+            F.one_hot(clas_tgt, num_classes=self.n_c + 1)[:, 1:]
+        )  # no loss for the first(background) class
 
         # classification loss
         clas_loss = self.focal_loss(clas_pred, clas_tgt) / torch.clamp(
